try to look up names sphinx reports as missing
it seems like it should be very simple to figure out the names
that sphinx does not find by just inpsecting the object itself,
but it's not clear if we can reliably look up the actual object
from what's in the reference node.   will try to come
back to this at some point.