Support module prefixes
This allows modulenames to be specified using absolute module
paths that are nonetheless succinct so that they are easy to
type and don't explode the verbosity of the documentation.  They
also provide a way of changing the location of top level modules
using configuration.  assuming no issues, this approach should ideally
be applied to all links throughout SQLAlchemy documentation so that
we no longer need to rely on sphinx's prefix "." idea which doesn't
really scale for something like SQLAlchemy.