host fonts on www.sqlalchemy.org
Google fonts are self-hosted so that there is no delay due to unavailable
network or firewalls within the @import tag, not just for loading the font
itself, which is resolvable with "font-display: swap", but for the css import
itself, which may still delay for a network timeout period before allowing
the page to load, particularly for static file documentation downloads.
Google does not offer stable font URLs.

Fixes: https://github.com/sqlalchemy/sqlalchemy/issues/5727