Vendor Sphinx viewcode from Sphinx 3.4.3
viewcode documents that it can import modules at any
time, so vendor the parts we need so that there is no
chance of an unexpected module import due to sphinx
code changes such as https://github.com/sphinx-doc/sphinx/issues/8681.

Reformat comment in docs.scss that seemed to be failing
for most recent pyscss release.