dont render for inherited attribute that wasn't found
sphinx can now send us attributes that aren't in
any cls.__dict__, like those which are only annotated.

for now, make sure we don't write the "inherits" indicator
if we didn't actually find the attribute on a superclass,
as this was writing it out for the base "object" class.

to fully implement rendering of "inherits" we may have
to look in __annotations__ but it's not clear if these
steps are necessary or useful right now.