attempt to show a note for annotated vs. nonannotaed code examples
im anticipating people being confused by annotations present
or not in different examples, so trying to insert a note.

I wanted to make it link to a glossary term but that's not
figured out yet.  use a transparent, faded background image
with careful scaling so it isn't too intrusive at least.

it looks good enough that it is likely more helpful than it is
distracting for now.